# Use Horner's method for evaluating polynomials