--- a/src/sss/encode.rs
+++ b/src/sss/encode.rs
@@ -3,14 +3,12 @@ use std::io;
 use std::io::prelude::*;
 
 /// evaluates a polynomial at x=1, 2, 3, ... n (inclusive)
-pub(crate) fn encode_secret_byte<W: Write>(src: &[u8], n: u8, w: &mut W) -> io::Result<()> {
+pub(crate) fn encode_secret_byte<W: Write>(src: &[u8], k: u8, n: u8, w: &mut W) -> io::Result<()> {
     for raw_x in 1..(u16::from(n) + 1) {
         let x = Gf256::from_byte(raw_x as u8);
-        let mut fac = Gf256::one();
-        let mut acc = Gf256::zero();
-        for &coeff in src.iter() {
-            acc += fac * Gf256::from_byte(coeff);
-            fac *= x;
+        let mut acc = Gf256::from_byte(src[(k - 1) as usize]);
+        for &coeff in src[..((k - 1) as usize)].iter().rev() {
+            acc = Gf256::from_byte(coeff) + acc * x;
         }
         w.write_all(&[acc.to_byte()])?;
     }