@nvesely 2018-03-15
Closing in favor of #50.